Move vallox base entity to separate module
