+++ homeassistant/components/vallox/date.py
@@ -12,9 +12,9 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import ValloxEntity
 from .const import DOMAIN
 from .coordinator import ValloxDataUpdateCoordinator
+from .entity import ValloxEntity
 
 
 class ValloxFilterChangeDateEntity(ValloxEntity, DateEntity):
