+++ homeassistant/components/vallox/switch.py
@@ -13,9 +13,9 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import ValloxEntity
 from .const import DOMAIN
 from .coordinator import ValloxDataUpdateCoordinator
+from .entity import ValloxEntity
 
 
 class ValloxSwitchEntity(ValloxEntity, SwitchEntity):
